requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

google-api-python-client==2.127.0
google-auth-oauthlib==1.2.0
halo==0.0.31
openai==1.23.6
pytz==2024.1
requests==2.31.0
tqdm==4.66.2
twilio==9.0.5
